Organizacion de componentes y carpetes en toda la app
