+++ frontend/.vite/deps/_metadata.json
@@ -1,0 +1,8 @@
+{
+  "hash": "8aee7317",
+  "configHash": "0eee6faa",
+  "lockfileHash": "e46c99cd",
+  "browserHash": "aead324f",
+  "optimized": {},
+  "chunks": {}
+}
