+++ frontend/src/components/AnalyticsResumeCard.tsx
