+++ frontend/src/components/Button.tsx
@@ -1,13 +1,3 @@
-export interface User {
-    id?: string;
-    name: string;
-    lastName: string;
-    email: string;
-    password: string;
-    birthDate: string;
-    avatar?: string;
-}
-
 type Props = {
     text: string;
     type: 'submit' | 'reset' | 'button' | undefined;
