+++ frontend/src/components/Card.tsx
