+++ frontend/src/components/DecisionFilters.tsx
@@ -1,5 +1,3 @@
-//import { useState } from 'react';
-
 const DecisionFilters = ({
     active,
     setActive,
@@ -7,7 +5,6 @@ const DecisionFilters = ({
     active: number;
     setActive: (index: number) => void;
 }) => {
-    //const [active, setActive] = useState<number>(1);
     return (
         <div className="inline-flex bg-gray-100 my-4 p-2 text-sm sm:text-lg items-center justify- rounded-lg">
             <ul className="flex flex-row font-medium items-center text-gray-800">
