+++ frontend/src/components/QuickStatsCard.tsx
