--- frontend/src/components/Register/BreadcrumbForm.tsx
@@ -1,9 +1,0 @@
-type Props = {
-    step: number;
-};
-
-const BreadcrumbForm = ({ step }: Props) => {
-    return <div>{step}</div>;
-};
-
-export default BreadcrumbForm;
