+++ frontend/src/components/SideNavBar.tsx
