+++ frontend/src/components/analytics/TinyBarChart.tsx
