+++ frontend/src/components/home/Hero.tsx
