+++ frontend/src/types/evaluation.types.ts
@@ -1,0 +1,8 @@
+export interface Evaluation {
+    id?: string;
+    decisionId: string;
+    score: number;
+    createdAt: string;
+    result?: string;
+    date?: string;
+}
