+++ frontend/src/types/recommendation.types.ts
@@ -1,0 +1,4 @@
+export interface Recommendation {
+    id: string;
+    decisionId: string;
+}
