+++ frontend/src/types/user.types.ts
@@ -1,0 +1,9 @@
+export interface User {
+    id?: string;
+    name: string;
+    lastName: string;
+    email: string;
+    password: string;
+    birthDate: string;
+    avatar?: string;
+}
